+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 - Can now assert that a resource object is one of multiple types using `assertTypeIs()`.
 - Can now assert that the `data` member of a document is `null`.
+- Can now assert that a resource object matches an expected structure.
 
 ### Fixed
 - Resource object type assertion caused a PHP error.